refactor: Update auth context import path and simplify user ID retrieval in chatroom components
